R6 class of test dataset for performance evaluation tools

TestDataB is a class that contains scores and label for performance evaluation tools. It provides necessary methods for benchmarking.

Format

An R6 class object.

Methods

get_tsname(): Get the dataset name.

get_scores(): Get a vector of scores.

get_labels(): Get a vector of labels.

get_fg(): Get a vector of positive scores.

get_bg(): Get a vector of negative scores.

get_fname(): Get a file name that contains scores and labels.

del_file(): Delete the file with scores and labels.

See also

Methods

Public methods

TestDataB$new()

TestDataB$get_tsname()

TestDataB$get_scores()

TestDataB$get_labels()

TestDataB$get_fg()

TestDataB$get_bg()

TestDataB$get_fname()

TestDataB$del_file()

TestDataB$print()

TestDataB$clone()

Method new()

Usage

TestDataB$new(scores = NULL, labels = NULL, tsname = NA)

Method get_tsname()

Usage

TestDataB$get_tsname()

Method get_scores()

Usage

TestDataB$get_scores()

Method get_labels()

Usage

TestDataB$get_labels()

Method get_fg()

Usage

TestDataB$get_fg()

Method get_bg()

Usage

TestDataB$get_bg()

Method get_fname()

Usage

TestDataB$get_fname()

Method del_file()

Usage

TestDataB$del_file()

Method print()

Usage

TestDataB$print(...)

Method clone()

The objects of this class are cloneable with this method.

Usage

TestDataB$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## Initialize with scores, labels, and a dataset name
testset <- TestDataB$new(c(0.1, 0.2, 0.3), c(0, 1, 1), "m1")
testset

#> 
#>     === Test dataset for prcbench functions ===
#> 
#>     Testset name:     m1 
#>     # of positives:   2 
#>     # of negatives:   1 
#>     Scores:           0.1 (min) 
#>                       0.2 (mean) 
#>                       0.3 (max) 
#>     Labels:           0 (neg), 1 (pos)
#>