R6 class of test dataset for Precision-Recall curve evaluation

TestDataC is a class that contains scores and label for performance evaluation tools. It provides necessary methods for curve evaluation.

TestDataC

Format

An R6 class object.

Methods

set_basepoints_x(x): Set pre-calculated recall values for curve evaluation

set_basepoints_y(y): Set pre-calculated precision values for curve evaluation

get_basepoints_x(): Get pre-calculated recall values for curve evaluation

get_basepoints_y(): Get pre-calculated precision values for curve evaluation

set_textpos_x(x): Set the x position for displaying the test result in a plot

set_textpos_y(y): Set the y position for displaying the test result in a plot

get_textpos_x(): Get the x position for displaying the test result in a plot

get_textpos_y(): Get the y position for displaying the test result in a plot

Following seven methods are inherited from TestDataB. See TestDataB for the method descriptions.

get_datname()

get_scores()

get_labels()

get_fg()

get_bg()

get_fname()

del_file()

See also

Examples

## Initialize with scores, labels, and a dataset name
testset <- TestDataC$new(c(0.1, 0.2), c(1, 0), "c4")
testset

#> 
#>     === Test dataset for prcbench functions ===
#> 
#>     Testset name:     c4 
#>     # of positives:   1 
#>     # of negatives:   1 
#>     Scores:           0.1 (min) 
#>                       0.15 (mean) 
#>                       0.2 (max) 
#>     Labels:           0 (neg), 1 (pos)
#>     Pre-calculated:   Yes
#>     # of base points: 1 
#>     Text position:    (0.75, 0.25)
#>     Text position2:   (0.75, 0.25)
#> 

## Set base points
testset$set_basepoints_x(c(0.13, 0.2))
testset$set_basepoints_y(c(0.5, 0.6))
testset

#> 
#>     === Test dataset for prcbench functions ===
#> 
#>     Testset name:     c4 
#>     # of positives:   1 
#>     # of negatives:   1 
#>     Scores:           0.1 (min) 
#>                       0.15 (mean) 
#>                       0.2 (max) 
#>     Labels:           0 (neg), 1 (pos)
#>     Pre-calculated:   Yes
#>     # of base points: 2 
#>     Text position:    (0.75, 0.25)
#>     Text position2:   (0.75, 0.25)
#>