R6 class of the AUCCalculator tool

ToolAUCCalculator is a wrapper class for the AUCCalculator tool, which is a Java library that provides calculations of ROC and Precision-Recall curves.

ToolAUCCalculator

Format

An R6 class object.

Inheritance

ToolIFBase

Methods

set_jarpath(jarpath)

It sets an AUCCalculator jar file.

jarpath

File path of the AUCCalculator jar file, e.g. "/path1/path2/auc2.jar".

Following nine methods are inherited from ToolIFBase. See ToolIFBase for the method descriptions.

call((testset, calc_auc, store_res)

get_toolname()

set_toolname(toolname)

get_setname()

set_setname(setname)

get_result()

get_x()

get_y()

get_auc()

See also

Examples

## Initialization
toolauccalc <- ToolAUCCalculator$new()

## Show object info
toolauccalc

#> 
#>     === Tool interface ===
#> 
#>     Tool name:            AUCCalculator 
#>     Calculate AUC score:  Yes
#>     Store results:        Yes
#>     Prediction performed: No
#>     Available methods:    call(testset, calc_auc, store_res)
#>                           get_toolname()
#>                           set_toolname(toolname)
#>                           get_setname()
#>                           set_setname(setname)
#>                           get_result()
#>                           get_x()
#>                           get_y()
#>                           get_auc()
#>                           set_jarpath(jarpath)
#>                           set_curvetype(curvetype)
#>                           set_auctype(auctype)
#>     Help file:            help("ToolAUCCalculator")

## create_toolset should be used for benchmarking and curve evaluation
toolauccalc2 <- create_toolset("AUCCalculator")