R6 class of the PerfMeas tool

ToolPerfMeas is a wrapper class for the PerfMeas tool, which is an R library that provides several performance measures.

ToolPerfMeas

Format

An R6 class object.

Inheritance

ToolIFBase

Methods

Following nine methods are inherited from ToolIFBase. See ToolIFBase for the method descriptions.

call(testset, calc_auc, store_res)

get_toolname()

set_toolname(toolname)

get_setname()

set_setname(setname)

get_result()

get_x()

get_y()

get_auc()

See also

This class is derived from ToolIFBase. create_toolset for creating a list of tools.

Examples

## Initialization
toolperf <- ToolPerfMeas$new()

## Show object info
toolperf

#>
#>     === Tool interface ===
#>
#>     Tool name:            PerfMeas
#>     Calculate AUC score:  Yes
#>     Store results:        Yes
#>     Prediction performed: No
#>     Available methods:    call(testset, calc_auc, store_res)
#>                           get_toolname()
#>                           set_toolname(toolname)
#>                           get_setname()
#>                           set_setname(setname)
#>                           get_result()
#>                           get_x()
#>                           get_y()
#>                           get_auc()
#>     Help file:            help("ToolPerfMeas")

## create_toolset should be used for benchmarking and curve evaluation
toolperf2 <- create_toolset("PerfMeas")