Base class of performance evaluation tools

ToolIFBase is an abstract class to provide a uniform interface for performance evaluation tools.

ToolIFBase

Format

An R6 class object

Methods

call(testset, calc_auc, store_res): It calls an actual tool to calculate Precision-Recall curves.

testset

R6 object generated by the create_testset function.

calc_auc

A Boolean value to specify whether the AUC score should be calculated.

store_res

A Boolean value to specify whether the calculated curve is retrieved and stored.

get_toolname(): Get the name of the tool.

set_toolname(toolname): Set the name of the tool.

get_setname(): Get the name of the tool set.

set_setname(setname): Set the name of the tool set.

get_result(): Get a list with curve values and the AUC score.

get_x(): Get calculated recall values.

get_y(): Get calculated precision values.

get_auc(): Get tne AUC score.
.

See also

ToolROCR, ToolAUCCalculator, ToolPerfMeas, ToolPRROC, and Toolprecrec are derived from this class. create_toolset for creating a list of tools.