ToolIFBase

Base class of performance evaluation tools.

Format

An R6 class object

Details

ToolIFBase is an abstract class to provide a uniform interface for performance evaluation tools.

See also

Methods

Public methods

ToolIFBase$new()

ToolIFBase$call()

ToolIFBase$get_toolname()

ToolIFBase$set_toolname()

ToolIFBase$get_setname()

ToolIFBase$set_setname()

ToolIFBase$get_result()

ToolIFBase$get_x()

ToolIFBase$get_y()

ToolIFBase$get_auc()

ToolIFBase$print()

ToolIFBase$clone()

Method new()

Default class initialization method.

Usage

ToolIFBase$new(...)

Arguments

...

set value for setname, calc_auc, store_res, x, y.

Method call()

It calls the tool to calculate precision-recall curves.

Usage

ToolIFBase$call(testset, calc_auc, store_res)

Arguments

testset

R6 object generated by the create_testset function.

calc_auc

A Boolean value to specify whether the AUC score should be calculated.

store_res

A Boolean value to specify whether the calculated curve is retrieved and stored.

Method get_toolname()

Get the name of the tool.

Usage

ToolIFBase$get_toolname()

Method set_toolname()

Set the name of the tool.

Usage

ToolIFBase$set_toolname(toolname)

Arguments

toolname

Name of the tool.

Method get_setname()

Get the name of the tool set.

Usage

ToolIFBase$get_setname()

Method set_setname()

Set the name of the tool set.

Usage

ToolIFBase$set_setname(setname)

Arguments

setname

Name of the tool set.

Method get_result()

Get a list with curve values and the AUC score.

Usage

ToolIFBase$get_result()

Method get_x()

Get calculated recall values.

Usage

ToolIFBase$get_x()

Method get_y()

Get calculated precision values.

Usage

ToolIFBase$get_y()

Method get_auc()

Get tne AUC score.

Usage

ToolIFBase$get_auc()

Method print()

Pretty print of the tool interface

Usage

ToolIFBase$print(...)

Arguments

...

Not used.

Method clone()

The objects of this class are cloneable with this method.

Usage

ToolIFBase$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.