R6 class of the PRROC tool

ToolPRROC is a wrapper class for the PRROC tool, which is an R library that provides calculations of ROC and Precision-Recall curves.

Format

An R6 class object.

Inheritance

ToolIFBase

Methods

set_curve(val)

A Boolean value to specify whether a Precision-Recall curve is calculated.

set_minStepSize(val)

A numeric value to specify the minimum step size between two intermediate points.

Following nine methods are inherited from ToolIFBase. See ToolIFBase for the method descriptions.

call(testset, calc_auc, store_res)

get_toolname()

set_toolname(toolname)

get_setname()

set_setname(setname)

get_result()

get_x()

get_y()

get_auc()

See also

Super class

prcbench::ToolIFBase -> ToolPRROC

Methods

Public methods

ToolPRROC$new()

ToolPRROC$set_curve()

ToolPRROC$set_minStepSize()

ToolPRROC$set_aucType()

ToolPRROC$clone()

Inherited methods

prcbench::ToolIFBase$call()

prcbench::ToolIFBase$get_auc()

prcbench::ToolIFBase$get_result()

prcbench::ToolIFBase$get_setname()

prcbench::ToolIFBase$get_toolname()

prcbench::ToolIFBase$get_x()

prcbench::ToolIFBase$get_y()

prcbench::ToolIFBase$print()

prcbench::ToolIFBase$set_setname()

prcbench::ToolIFBase$set_toolname()

Method new()

Usage

ToolPRROC$new(...)

Method set_curve()

Usage

ToolPRROC$set_curve(val)

Method set_minStepSize()

Usage

ToolPRROC$set_minStepSize(val)

Method set_aucType()

Usage

ToolPRROC$set_aucType(val)

Method clone()

The objects of this class are cloneable with this method.

Usage

ToolPRROC$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## Initialization
toolprroc <- ToolPRROC$new()

## Show object info
toolprroc

#> 
#>     === Tool interface ===
#> 
#>     Tool name:            PRROC 
#>     Calculate AUC score:  Yes
#>     Store results:        Yes
#>     Prediction performed: No
#>     Available methods:    call(testset, calc_auc, store_res)
#>                           get_toolname()
#>                           set_toolname(toolname)
#>                           get_setname()
#>                           set_setname(setname)
#>                           get_result()
#>                           get_x()
#>                           get_y()
#>                           get_auc()
#>                           set_curve(val)
#>                           set_minStepSize(val)
#>                           set_aucType(val)
#>     Help file:            help("ToolPRROC")

## create_toolset should be used for benchmarking and curve evaluation
toolprroc2 <- create_toolset("PRROC")