ToolPRROC

R6 class of the PRROC tool

Format

An R6 class object.

Details

ToolPRROC is a wrapper class for the PRROC tool, which is an R library that provides calculations of ROC and Precision-Recall curves.

See also

Super class

prcbench::ToolIFBase -> ToolPRROC

Methods

Public methods

ToolPRROC$new()

ToolPRROC$set_curve()

ToolPRROC$set_minStepSize()

ToolPRROC$set_aucType()

ToolPRROC$clone()

Inherited methods

prcbench::ToolIFBase$call()

prcbench::ToolIFBase$get_auc()

prcbench::ToolIFBase$get_result()

prcbench::ToolIFBase$get_setname()

prcbench::ToolIFBase$get_toolname()

prcbench::ToolIFBase$get_x()

prcbench::ToolIFBase$get_y()

prcbench::ToolIFBase$print()

prcbench::ToolIFBase$set_setname()

prcbench::ToolIFBase$set_toolname()

Method new()

Default class initialization method.

Usage

ToolPRROC$new(...)

Arguments

...

set value for curve, minStepSize, aucType.

Method set_curve()

A Boolean value to specify whether precision-recall curve is calculated.

Usage

ToolPRROC$set_curve(val)

Arguments

val

TRUE: calculate, FALSE: not calculate.

Method set_minStepSize()

A numeric value to specify the minimum step size between two intermediate points.

Usage

ToolPRROC$set_minStepSize(val)

Arguments

val

Step size between two points.

Method set_aucType()

Set the AUC calculation method

Usage

ToolPRROC$set_aucType(val)

Arguments

val

1: integral, 2: Davis Goadrich

Method clone()

The objects of this class are cloneable with this method.

Usage

ToolPRROC$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## Initialization
toolprroc <- ToolPRROC$new()

## Show object info
toolprroc
#> 
#>     === Tool interface ===
#> 
#>     Tool name:            PRROC 
#>     Calculate AUC score:  Yes
#>     Store results:        Yes
#>     Prediction performed: No
#>     Available methods:    call(testset, calc_auc, store_res)
#>                           get_toolname()
#>                           set_toolname(toolname)
#>                           get_setname()
#>                           set_setname(setname)
#>                           get_result()
#>                           get_x()
#>                           get_y()
#>                           get_auc()
#>                           set_curve(val)
#>                           set_minStepSize(val)
#>                           set_aucType(val)
#>     Help file:            help("ToolPRROC")

## create_toolset should be used for benchmarking and curve evaluation
toolprroc2 <- create_toolset("PRROC")