ToolROCR

R6 class of the ROCR tool

Format

An R6 class object.

Details

ToolROCR is a wrapper class for the ROCR tool, which is an R library that provides calculations of various performance evaluation measures.

See also

Super class

prcbench::ToolIFBase -> ToolROCR

Methods

Public methods

ToolROCR$clone()

Inherited methods

prcbench::ToolIFBase$call()

prcbench::ToolIFBase$get_auc()

prcbench::ToolIFBase$get_result()

prcbench::ToolIFBase$get_setname()

prcbench::ToolIFBase$get_toolname()

prcbench::ToolIFBase$get_x()

prcbench::ToolIFBase$get_y()

prcbench::ToolIFBase$initialize()

prcbench::ToolIFBase$print()

prcbench::ToolIFBase$set_setname()

prcbench::ToolIFBase$set_toolname()

Method clone()

The objects of this class are cloneable with this method.

Usage

ToolROCR$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## Initialization
toolrocr <- ToolROCR$new()

## Show object info
toolrocr

#> 
#>     === Tool interface ===
#> 
#>     Tool name:            ROCR 
#>     Calculate AUC score:  Yes
#>     Store results:        Yes
#>     Prediction performed: No
#>     Available methods:    call(testset, calc_auc, store_res)
#>                           get_toolname()
#>                           set_toolname(toolname)
#>                           get_setname()
#>                           set_setname(setname)
#>                           get_result()
#>                           get_x()
#>                           get_y()
#>                           get_auc()
#>     Help file:            help("ToolROCR")

## create_toolset should be used for benchmarking and curve evaluation
toolrocr2 <- create_toolset("ROCR")