Toolprecrec

R6 class of the precrec tool

Format

An R6 class object.

Details

Toolprecrec is a wrapper class for the precrec tool, which is an R library that provides calculations of ROC and Precision-Recall curves.

See also

Super class

prcbench::ToolIFBase -> Toolprecrec

Methods

Public methods

Toolprecrec$new()

Toolprecrec$set_x_bins()

Toolprecrec$clone()

Inherited methods

prcbench::ToolIFBase$call()

prcbench::ToolIFBase$get_auc()

prcbench::ToolIFBase$get_result()

prcbench::ToolIFBase$get_setname()

prcbench::ToolIFBase$get_toolname()

prcbench::ToolIFBase$get_x()

prcbench::ToolIFBase$get_y()

prcbench::ToolIFBase$print()

prcbench::ToolIFBase$set_setname()

prcbench::ToolIFBase$set_toolname()

Method new()

Default class initialization method.

Usage

Toolprecrec$new(...)

Arguments

...

set value for x_bins.

Method set_x_bins()

Set the number of supporting points as the number of bins.

Usage

Toolprecrec$set_x_bins(x_bins)

Arguments

x_bins

set value for x_bins.

Method clone()

The objects of this class are cloneable with this method.

Usage

Toolprecrec$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## Initialization
toolprecrec <- Toolprecrec$new()

## Show object info
toolprecrec

#> 
#>     === Tool interface ===
#> 
#>     Tool name:            precrec 
#>     Calculate AUC score:  Yes
#>     Store results:        Yes
#>     Prediction performed: No
#>     Available methods:    call(testset, calc_auc, store_res)
#>                           get_toolname()
#>                           set_toolname(toolname)
#>                           get_setname()
#>                           set_setname(setname)
#>                           get_result()
#>                           get_x()
#>                           get_y()
#>                           get_auc()
#>                           set_x_bins(x_bins)
#>     Help file:            help("Toolprecrec")

## create_toolset should be used for benchmarking and curve evaluation
toolprecrec2 <- create_toolset("precrec")