Run microbenchmark with specified tools and test sets

The run_benchmark function runs microbenchmark for specified tools and test datasets

run_benchmark(testset, toolset, times = 5, unit = "ms",
  use_sys_time = FALSE)

Arguments

testset
A character vector to specify a test set generated by create_testset.

toolset
A character vector to specify a tool set generated by create_toolset.

times
The number of iteration used in microbenchmark.

unit
A single string to specify the unit used in summary.microbenchmark.

use_sys_time
A Boolean value to specify system.time is used instead of summary.microbenchmark.

Value

A data frame of microbenchmark results with additional columns.

See also

create_testset to generate a test dataset. create_toolset to generate a tool set. microbenchmark for benchmarking details.

Examples

## Benchmarking for b10 and i10 test sets and crv5, auc5, and def5 tool sets
testset <- create_testset("bench", c("b10", "i10"))
toolset <- create_toolset(set_names = "def5")
res1 <- run_benchmark(testset, toolset)
res1

#>    testset toolset      toolname  min   lq  mean median    uq   max neval
#> 1      b10    def5          ROCR 2.56 2.62  7.83   2.78  3.13 28.08     5
#> 2      b10    def5 AUCCalculator 7.66 8.55 21.69   9.60 10.13 72.53     5
#> 3      b10    def5      PerfMeas 0.13 0.13  2.32   0.21  0.43 10.72     5
#> 4      b10    def5         PRROC 1.05 1.08  1.25   1.17  1.43  1.53     5
#> 5      b10    def5       precrec 8.13 8.77 10.23  10.60 11.77 11.90     5
#> 6      i10    def5          ROCR 2.62 2.67  2.77   2.73  2.73  3.12     5
#> 7      i10    def5 AUCCalculator 8.23 8.25  8.46   8.26  8.72  8.85     5
#> 8      i10    def5      PerfMeas 0.13 0.13  0.16   0.13  0.14  0.29     5
#> 9      i10    def5         PRROC 0.59 0.67  0.74   0.71  0.74  1.01     5
#> 10     i10    def5       precrec 8.11 8.18  8.26   8.25  8.31  8.45     5