Run microbenchmark with specified tools and test sets

The run_benchmark function runs microbenchmark for specified tools and test datasets

run_benchmark(testset, toolset, times = 5, unit = "ms", use_sys_time = FALSE)

Arguments

testset
A character vector to specify a test set generated by create_testset.

toolset
A character vector to specify a tool set generated by create_toolset.

times
The number of iteration used in microbenchmark.

unit
A single string to specify the unit used in summary.microbenchmark.

use_sys_time
A Boolean value to specify system.time is used instead of summary.microbenchmark.

Value

A data frame of microbenchmark results with additional columns.

See also

Examples

# \dontrun{
## Benchmarking for b10 and i10 test sets and crv5, auc5, and def5 tool sets
testset <- create_testset("bench", c("b10", "i10"))
toolset <- create_toolset(set_names = "def5")
res1 <- run_benchmark(testset, toolset)
res1

#>    testset toolset      toolname   min    lq mean median   uq   max neval
#> 1      b10    def5 AUCCalculator 5.906 5.919 7.75  7.243 7.30 12.40     5
#> 2      b10    def5      PerfMeas 0.084 0.086 0.15  0.094 0.11  0.36     5
#> 3      b10    def5       precrec 5.299 5.369 5.89  5.710 6.24  6.82     5
#> 4      b10    def5         PRROC 0.455 0.458 0.50  0.458 0.48  0.67     5
#> 5      b10    def5          ROCR 2.575 2.578 3.64  3.807 4.34  4.89     5
#> 6      i10    def5 AUCCalculator 6.250 6.946 8.07  7.206 7.86 12.09     5
#> 7      i10    def5      PerfMeas 0.155 0.166 0.20  0.178 0.19  0.29     5
#> 8      i10    def5       precrec 6.032 6.468 7.20  6.590 7.62  9.31     5
#> 9      i10    def5         PRROC 0.439 0.441 0.49  0.456 0.49  0.65     5
#> 10     i10    def5          ROCR 2.160 2.246 2.49  2.416 2.60  3.02     5

# }